pyproject.toml
[project]
name = "pi-hailo-ml"
version = "0.1.0"
description = "Viam modular resource for using CV models on Raspberry Pi AI Kit/HAT hardware"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy==1.26.4",
    "opencv-python>=4.10.0.84",
    "pillow>=11.1.0",
    "pydantic>=2.10.3",
    "typing-extensions>=4.12.2",
    "viam-sdk==0.37.0",
]
